Extract logic to EmbedComponent to separate concerns
